Move practical date formating functions into utils so it's available for both modules
