+++ Source/SelfService/Web/applications/microservice/helpers/index.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 export { capitalize, removeStringPattern, removeStringPatternAndCapitalize, getRuntimeNumberFromString } from './strings';
-export { formatTime, formatStartingDate } from './dates';
+export { formatTime, formatStartingDate } from '../../../utils/dates';
 export { DownloadLogs } from './downloads';
 export { formatBytes } from './numbers';
 export { isAlphaNumeric } from './regex';
